feat : 모임 서비스

+++ backend/src/main/kotlin/com/finance/backend/group/request/DuesPayReq.kt
@@ -6,7 +6,7 @@ data class DuesPayReq(
         @JsonProperty("dues_id")
         val duesId : Long,
         @JsonProperty("dues_val")
-        val duesVal : Long,
+        val duesVal : Int,
         @JsonProperty("ac_no")
         val acNo : String
         )
